refactor: split persistence auto-save debounce into own module
